final cleanup/simplification before TESTING
